lib: Typechecking minmax
Add typechecking to the min/max macros and make their locals truly
unique-ish to reduce the risk of shadowing.

v2: small bug fix, write also height coordinate on rotation
    test. (jheikkil)
v3: Fix up a couple of other max/max_t instances (Ashutosh)

+++ lib/i915/intel_memory_region.h
@@ -58,7 +58,7 @@ uint32_t gem_get_batch_size(int fd, uint8_t mem_region_type);
 
 struct drm_i915_query_memory_regions *gem_get_query_memory_regions(int fd);
 
-uint8_t gem_get_lmem_region_count(int fd);
+unsigned int gem_get_lmem_region_count(int fd);
 
 bool gem_has_lmem(int fd);
